pyproject.toml
[project]
name = "hedgy-options"
version = "0.1.0"
description = ""
authors = [
    {name = "Tyler Pruitt",email = "tylerp@creatrify.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "robin-stocks (>=3.4.0,<4.0.0)",
    "scipy (>=1.17.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "hedgy_options", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
